LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/moving-between-documents/move-back-iframe-success-external-classic-expected.txt

CONSOLE MESSAGE: exec
 

PASS Sanity check around top-level Window 
FAIL Eval: Move success external classic script to iframe move-back assert_false: The script must not have executed in source window expected false got "executed"
FAIL <script> load: Move success external classic script to iframe move-back assert_unreached: Script load event fired unexpectedly Reached unreachable code
PASS <script> error: Move success external classic script to iframe move-back 
PASS window error: Move success external classic script to iframe move-back 
